Current CAN_Init Class and Prototype Filter_Config Class
Created a CAN_Init Class (which is in progress) and a prototype transmit class. Considering making a parent class to facilitate multiple CAN Instances. If there are multiple CAN instances, the parent class can manage which CAN is working and which one isn't, allowing for easy switching in case the master fails.